pyproject.toml
[project]
name = "prompt2post"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gTTS>=2.5.1",
    "python-dotenv==1.1.1",
    "pytrends==4.9.2",
    "requests==2.32.4",
    "streamlit==1.48.1",
]

